Rewrite the lexer to be more functional
Rewrite the lexer component to be much more functional and extensible.
Indentation parsing uses a stack to track previous levels and the main loop
does as little work as possible before handing off to other functions.

Each 'type' of token has its own function that deals with reading it from the
input and placing it on the queue. This means the lexer can read multiple
tokens in a single invocation if possible, which then get placed on the queue
for reading.