Allow for blank lines in the source file
Allow for blank lines in the source file to allow nicer code to be input.
Update a Suite to be a list of Lines, which can be either a newline character
or a Statement. This should be checked against the official Python grammar to
see how it handles this.

Lexer will now read indentation and check whether we are at a newline
character, thus indicating we are on a blank line. It will read a newline
token, and then recursively call itself to read more blank lines.