Add output to a file with the same name as input
Implement output to a file with the same name as the input file. For example,
running the program on max.py will create a new file called max.c with the
generated C code. If the file already exists, the program will display an error
message, otherwise it will write the output code to the file.