Add a lib.rs to allow for tests in crate/tests
Add a lib.rs file in src/ to allow for tests to be placed in the tests/
directory. This allows for better separation using cargo test when running the
tests, and allows them to be run independently more easily.

Lexer tests are unit tests and thus have been placed in src/lexer.rs, but the
parser and codegen tests are integration tests and thus are placed in
tests/parser_tests.rs and tests/codegen_tests.rs respectively.