Fix len() inference
Fix len() inference to not assume a List[int] and instead just assume a List.
This caused a panic as it was generating a ListDisplay, which is unreachable.
Thus, `check_list_display` needs to generate if it is a List, not if it also
knows the element type. The tests also need to be updated to reflect this.