Update to actually use the proper frees
Update the compiler to use the correct free() functions so that it doesn't
completely leak memory everywhere. Previously, it would just free the pointer
to the list structure, whereas now it will free that pointer along with the
data pointer. This also means that `stdlib.h` is no longer needed as an
include.