Add a check into insert_shallow_inferred_type
Add a check into insert_shallow_inferred_type to ensure that we do not
overwrite the element type of a variable using this function. This can occur in
the following snippet:

```python
def main():
	integers: List[int] = []
	len(integers)
```

Here, we would use the type hint in the initialisation to infer a type of
`VariableType::List { elements: VariableType::Integer }` which would get
inserted into the symbol table.

The compiler would then progress to the `len` call, and use the
`insert_shallow_inferred_type` to insert the `VariableType::List { elements:
None }` type into the symbol table.

If we don't know what is inside `integers`, this is the correct thing to do as
we at least now know it is a `VariableType::List` and can perform operations
with this. However, otherwise it can cause information loss for the compiler.

The `built_in_len_function` shouldn't have been changed originally as the type
as already known, so this was an oversight and has been reverted.